Move basic logic into own file
